+++ src/service/reserve.ts
@@ -1,7 +1,7 @@
 /**
  * 予約サービス
  */
-import * as cinerinoapi from '@cinerino/api-nodejs-client';
+import * as cinerinoapi from '@cinerino/sdk';
 import * as moment from 'moment';
 
 import * as factory from '@tokyotower/factory';
@@ -23,7 +23,7 @@ export function onReservationStatusChanged(
         const reservation = params;
 
         // 余分確保分を除く
-        let extraProperty: factory.propertyValue.IPropertyValue<string> | undefined;
+        let extraProperty: factory.chevre.propertyValue.IPropertyValue<string> | undefined;
         if (reservation.additionalProperty !== undefined) {
             extraProperty = reservation.additionalProperty.find((p) => p.name === 'extra');
         }